[Feat] 안 챙긴 짐 팝업 띄우기 기능 추가

+++ Packit/DTO/TripProgressResDto.swift
@@ -10,5 +10,5 @@ import Foundation
 struct TripProgressResDto: Codable {
     let totalCategoryCount: Int
     let completedCategoryCount: Int
-    let progressRate: Float
+    let progressRate: Double
 }
